Pasando cluster como clase independiente
